pyproject.toml
[project]
name = "Baseball_Manager"  
version = "0.1.0"           # Versión inicial del proyecto
dependencies = [
    "asgiref", #para poder instalar manualmente la ultima versión de django
    "chardet",
    "sqlparse", #lo mismo que el de arriba, si se instalada directo con pip no hace falta pensarlos porque se instalan automáticamente
    "django",
    "djangorestframework",
    "django-cors-headers",
    "factory-boy",
    "faker",
    "psycopg2-binary",
    "Pillow",                
    "django-extensions",     
    "python-decouple",
    "python-dateutil",
    "pytz",
    "reportlab", 
    "six",
    "typing-extensions",
]
